% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gfiConfInt}
\alias{gfiConfInt}
\title{Fiducial confidence intervals for ultrahigh-dimensional regression}
\usage{
gfiConfInt(gfi, conf = 0.95)
}
\arguments{
\item{gfi}{an output of \code{\link{gfiUltra}}}

\item{conf}{confidence level}
}
\value{
The confidence intervals in a matrix.
}
\description{
Fiducial confidence intervals of the selected parameters of a
  ultrahigh-dimensional regression.
}
\seealso{
\code{\link{gfiEstimates}}
}
