% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gfiEstimates}
\alias{gfiEstimates}
\title{Fiducial estimates for ultrahigh-dimensional regression}
\usage{
gfiEstimates(gfi)
}
\arguments{
\item{gfi}{an output of \code{\link{gfiUltra}}}
}
\value{
The estimates in a matrix.
}
\description{
Fiducial estimates of the selected parameters of a
  ultrahigh-dimensional regression.
}
\seealso{
\code{\link{gfiConfInt}}
}
