% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gfiUltra}
\alias{gfiUltra}
\title{Title}
\usage{
gfiUltra(formula, data, nsims = 5L, gamma = 1, ...)
}
\arguments{
\item{formula}{xxx}

\item{data}{xxx}

\item{gamma}{xxx}

\item{...}{named arguments passed to \code{\link[SIS:SIS]{SIS}}}
}
\value{
xxx
}
\description{
xxx
}
\examples{
xxx
}
